Extract HeaderContent
